Moved chomsky text to an own file.
